platform/infrastucture/requirements.txt
pulumi>=3.0.0,<4.0.0
pulumi-aws>=6.0.2,<7.0.0

pulumi_aws~=6.47.0
selenium~=4.24.0
pulumi_docker~=4.5.5
boto3~=1.35.35
attrs~=23.2.0
botocore~=1.35.35
pyjwt~=2.9.0
